requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
cachetools
certifi
charset-normalizer
deprecation
exceptiongroup
filelock
frozenlist
fsspec
google-auth
google-genai
h11
hf-xet
httpcore
httpx
huggingface-hub
idna
Jinja2
joblib
lancedb
MarkupSafe
mpmath
multidict
networkx
numpy
overrides
packaging
pandas
pillow
propcache
psutil
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
python-dateutil
pytz
PyYAML
regex
requests
rsa
safetensors
scikit-learn
scipy
sentence-transformers
six
sniffio
sympy
threadpoolctl
tokenizers
torch
torch-geometric
tqdm
transformers
typing-inspection
typing_extensions
tzdata
urllib3
websockets
yarl
